feature/cocktail-sort: Modules seperated.
