+++ app/views/layout/footer.php
@@ -5,7 +5,7 @@
 </footer>
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 <script src="https://code.jquery.com/ui/1.13.1/jquery-ui.min.js"></script>
-<script type="module" src="/assets/js/modules/main.js"></script>
+<script type="module" src="/assets/js/main.js"></script>
 
 </body>
 </html>
